+++ CHANGELOG.md
@@ -16,6 +16,9 @@ Don't forget to remove deprecated code on each major release!
 
 - Nothing (yet)!
 
+### Added
+- Added support for custom metadata writing and validation during operations via `DBBACKUP_CUSTOM_METADATA_LOADER` and `DBBACKUP_CUSTOM_METADATA_VALIDATOR` settings.
+
 ## [5.1.2] - 2026-01-14
 
 ### Fixed